backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2
reportlab
flask-cors